+++ crystals/tests/test_parsers.py
@@ -296,10 +296,9 @@ def test_pwscf_crystal_instance(file, expected_spg):
     assert crystal.symmetry(symprec=1e-1)["international_number"] == expected_spg
 
 
-@pytest.mark.parametrize("poscar", [
-    "issue5_MgSiO3.cartesian.POSCAR",
-    "issue5_MgSiO3.direct.POSCAR"
-])
+@pytest.mark.parametrize(
+    "poscar", ["issue5_MgSiO3.cartesian.POSCAR", "issue5_MgSiO3.direct.POSCAR"]
+)
 def test_pwscf_lattice_vectors_alat(poscar):
     """Test POSCAR parser against CIF parser"""
     c0 = Crystal.from_cif(Path(__file__).parent / "data" / "issue5_MgSiO3.cif")